refactor(client): extract AnnotationCard, FilterSelect, useReviewKeyboard from SidePanel
SidePanel.tsx drops from 732 to ~490 lines. Extracted components
wrapped with React.memo for list rendering performance.
