+++ src/client/panels/FilterSelect.tsx
@@ -1,0 +1,29 @@
+export interface FilterSelectProps {
+  value: string;
+  onChange: (v: string) => void;
+  options: Array<{ value: string; label: string }>;
+}
+
+export function FilterSelect({ value, onChange, options }: FilterSelectProps) {
+  return (
+    <select
+      value={value}
+      onChange={(e) => onChange(e.target.value)}
+      style={{
+        padding: "2px 4px",
+        fontSize: "11px",
+        border: "1px solid #e5e7eb",
+        borderRadius: "3px",
+        background: "#fff",
+        color: "#374151",
+        cursor: "pointer",
+      }}
+    >
+      {options.map((o) => (
+        <option key={o.value} value={o.value}>
+          {o.label}
+        </option>
+      ))}
+    </select>
+  );
+}
